Add primitive types for 7, 15, 31, and 63 bit unsigned integers.
Adding these primitive types has two benefits.  Firstly, unsigned integers that
narrowly exceed a particular unsigned primitive type previously got classified
as the next larger signed integer type.  With this change, they are classified
as an unsigned integer type and the sign bit is not lost.  The other benefit is
that with this change, there is now a 1:1 correspondence between primitive
types, and SBCL's unboxed array element types.