Handle if-statements more carefully when deriving specializers.
The then or else branch of an if statement may be unreachable at run time, but
potentially reachable during compile time.  This may result in a type error
being signaled when processing that branch, although the code is correct.  This
problem can be addressed in two ways:

1. Optimistically assume that erroneous branches are unreachable.

2. Conservatively give up type inference for the entire expression.

For now, we err on the side of caution and use the second option.